+++ CHANGELOG.md
@@ -16,11 +16,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added microgrid windturbine actions to myems-api and myems-admin
 - added get all associated objects of microgrid to myems-api
 - added new access_control procedure to mymes-api
-- added access_contorl to shopfloor GET actions in myems-api and myems-admin
-- added access_contorl to shopfloor reports actions in myems-api
 - added svg, inverters and converters to microgrid in system database
 - added svg to microgrid in myems-api and myems-admin
 - added API Key to database, myems-api and myems-admin
+- added access control to shopfloor in myems-api, myems-admin and myems-web
+- added access control to store in myems-api, myems-admin and myems-web
 - 
 ### Changed
 - renamed access_control to admin_control in myems-api